+++ .env.example
@@ -13,6 +13,12 @@ SUPABASE_ANON_PUBLIC="{ANON_PUBLIC}"
 SUPABASE_SERVICE_ROLE="{SERVICE_ROLE}"
 SUPABASE_URL="https://{YOUR_INSTANCE_NAME}.supabase.co"
 SERVER_URL="http://localhost:3000"
+
+# Used for shortening URLs of QR codes. Optional
+# If present, the QR code will be generated with a shortened URL
+# URL_SHORTENER="https://hey.lo/"
+
+
 APP_NAME="Shelf"
 
 # Used for generating cuid with lowered chance of collision. Optional
