+++ app/modules/qr/utils.server.ts
@@ -1,12 +1,63 @@
 import type { Organization, Qr, User } from "@prisma/client";
+import QRCode from "qrcode-generator";
+import { SERVER_URL, URL_SHORTENER } from "~/utils/env";
+import type { ErrorLabel } from "~/utils/error";
 import { isLikeShelfError, ShelfError } from "~/utils/error";
+import { gifToPng } from "~/utils/gif-to-png";
 // eslint-disable-next-line import/no-cycle
-import {
-  createQr,
-  generateCode,
-  getQrByAssetId,
-  getQrByKitId,
-} from "./service.server";
+import { createQr, getQrByAssetId, getQrByKitId } from "./service.server";
+
+const label: ErrorLabel = "QR";
+
+export function getQrBaseUrl() {
+  return URL_SHORTENER ? `${URL_SHORTENER}/qr` : `${SERVER_URL}/qr`;
+}
+
+export async function generateCode({
+  version,
+  errorCorrection,
+  qr,
+  size,
+}: {
+  version: TypeNumber;
+  errorCorrection: ErrorCorrectionLevel;
+  qr: Qr;
+  size: "cable" | "small" | "medium" | "large";
+}) {
+  const baseUrl = getQrBaseUrl();
+
+  try {
+    const code = QRCode(version, errorCorrection);
+    code.addData(`${baseUrl}/${qr.id}`);
+    code.make();
+
+    /** We use a margin of 0 because we handle this using canvas in the client */
+    const sizes = {
+      cable: [1, 0], // 29px => 0.8cm(0.77)
+      small: [2, 0], // 58px => 1.5cm(1.53)
+      medium: [4, 0], // 116px => 3.1cm(3.07)
+      large: [6, 0], // 174px => 4.7cm(4.6)
+    };
+    const src = await gifToPng(code.createDataURL(...sizes[size]));
+
+    return {
+      sizes,
+      code: {
+        size: size,
+        src,
+        id: qr.id,
+      },
+    };
+  } catch (cause) {
+    throw new ShelfError({
+      cause,
+      message:
+        "Something went wrong while generating the QR code. Please try again or contact support.",
+      additionalData: { version, errorCorrection, qr, size },
+      label,
+    });
+  }
+}
 
 export async function generateQrObj({
   kitId,
