+++ app/utils/env.ts
@@ -25,6 +25,7 @@ declare global {
       SUPABASE_URL: string;
       SUPABASE_SERVICE_ROLE: string;
       SERVER_URL: string;
+      URL_SHORTENER: string;
       SUPABASE_ANON_PUBLIC: string;
       SESSION_SECRET: string;
       MAPTILER_TOKEN: string;
@@ -106,6 +107,10 @@ export const SUPABASE_SERVICE_ROLE = getEnv("SUPABASE_SERVICE_ROLE");
 export const INVITE_TOKEN_SECRET = getEnv("INVITE_TOKEN_SECRET", {
   isSecret: true,
 });
+export const URL_SHORTENER = getEnv("URL_SHORTENER", {
+  isRequired: false,
+});
+
 export const SESSION_SECRET = getEnv("SESSION_SECRET");
 export const FINGERPRINT = getEnv("FINGERPRINT", {
   isSecret: true,
