+++ app/utils/zip-qr-codes.ts
@@ -2,13 +2,15 @@ import type { Qr } from "@prisma/client";
 import { ErrorCorrection } from "@prisma/client";
 import JSZip from "jszip";
 import QRCode from "qrcode-generator";
+import { getQrBaseUrl } from "~/modules/qr/utils.server";
 
 export async function createQrCodesZip(codes: Qr[]) {
   const zip = new JSZip();
+  const baseUrl = getQrBaseUrl();
 
   codes.forEach((c) => {
     const code = QRCode(0, ErrorCorrection["L"]);
-    code.addData(`${process.env.SERVER_URL}/qr/${c.id}`);
+    code.addData(`${baseUrl}/${c.id}`);
     code.make();
     const svg = code.createSvgTag({
       cellSize: 3,
